requirements_base.txt
otree-core>=1.4.5
# Django==1.8.8 # for heroku, needs to be explicitly in requirements file
pycryptodome
# xlrd
# xlsxwriter
# numpy